fixes to register flow
