+++ src/navigations/index.tsx
@@ -24,10 +24,9 @@ import {
   MailDetailsScreen,
   PrivacyScreen,
   ReferFriendsScreen,
-  Register1Screen,
-  Register2Screen,
-  Register3Screen,
-  Register4Screen,
+  RegisterCredsScreen,
+  RegisterPersonalScreen,
+  RegisterAddressScreen,
   ReviewLetterScreen,
   ReviewPostcardScreen,
   ReviewContactScreen,
@@ -43,7 +42,7 @@ import { AppState } from '@store/types';
 import { AuthInfo, UserState } from '@store/User/UserTypes';
 import { navigationRef, navigate } from '@notifications';
 import { Notif } from '@store/Notif/NotifTypes';
-import { SupportFAQTypes, DeliveryReportTypes, Category } from 'types';
+import { SupportFAQTypes, DeliveryReportTypes, Category, Image } from 'types';
 import Topbar, {
   setTitle,
   topbarRef,
@@ -63,30 +62,22 @@ export type AuthStackParamList = {
   Login: undefined;
   Terms: undefined;
   Privacy: undefined;
-  Register1: undefined;
-  Register2: { firstName: string; lastName: string; referrer: string };
-  Register3: {
-    firstName: string;
-    lastName: string;
-    referrer: string;
+  RegisterCreds: undefined;
+  RegisterPersonal: {
     email: string;
     password: string;
     passwordConfirmation: string;
     remember: boolean;
   };
-  Register4: {
-    firstName: string;
-    lastName: string;
-    referrer: string;
+  RegisterAddress: {
     email: string;
     password: string;
     passwordConfirmation: string;
     remember: boolean;
-    address1: string;
-    address2: string;
-    city: string;
-    phyState: string;
-    postal: string;
+    firstName: string;
+    lastName: string;
+    referrer: string;
+    image: Image | undefined;
   };
 };
 
@@ -132,10 +123,9 @@ const mapRouteNameToDetails: Record<string, RouteDetails> = {
   Login: { title: i18n.t('Screens.login'), profile: false },
   Terms: { title: i18n.t('Screens.termsOfService'), profile: false },
   Privacy: { title: i18n.t('Screens.privacyPolicy'), profile: false },
-  Register1: { title: i18n.t('Screens.register'), profile: false },
-  Register2: { title: i18n.t('Screens.register'), profile: false },
-  Register3: { title: i18n.t('Screens.register'), profile: false },
-  Register4: { title: i18n.t('Screens.register'), profile: false },
+  RegisterCreds: { title: i18n.t('Screens.register'), profile: false },
+  RegisterPersonal: { title: i18n.t('Screens.register'), profile: false },
+  RegisterAddress: { title: i18n.t('Screens.register'), profile: false },
   AddManually: { title: i18n.t('Screens.addManually'), profile: false },
   ChooseCategory: { title: i18n.t('Screens.compose'), profile: false },
   ChooseOption: { title: i18n.t('Screens.compose'), profile: false },
@@ -263,10 +253,15 @@ const NavigatorBase: React.FC<Props> = (props: Props) => {
         <Stack.Screen name="Login" component={LoginScreen} />
         <Stack.Screen name="Terms" component={TermsScreen} />
         <Stack.Screen name="Privacy" component={PrivacyScreen} />
-        <Stack.Screen name="Register1" component={Register1Screen} />
-        <Stack.Screen name="Register2" component={Register2Screen} />
-        <Stack.Screen name="Register3" component={Register3Screen} />
-        <Stack.Screen name="Register4" component={Register4Screen} />
+        <Stack.Screen name="RegisterCreds" component={RegisterCredsScreen} />
+        <Stack.Screen
+          name="RegisterPersonal"
+          component={RegisterPersonalScreen}
+        />
+        <Stack.Screen
+          name="RegisterAddress"
+          component={RegisterAddressScreen}
+        />
       </>
     );
   } else {
