+++ src/views/Onboarding/RegisterAddress.react.tsx
@@ -9,28 +9,38 @@ import {
 } from 'react-native';
 import { StackNavigationProp } from '@react-navigation/stack';
 import { AuthStackParamList } from '@navigations';
-import { Button, Input } from '@components';
+import { Input } from '@components';
 import i18n from '@i18n';
 import { Typography } from '@styles';
-import { Validation, STATES_DROPDOWN, sleep } from '@utils';
+import { Validation, STATES_DROPDOWN, sleep, hoursTill8Tomorrow } from '@utils';
+import { Image } from 'types';
+import * as Segment from 'expo-analytics-segment';
+import { UserRegisterInfo } from '@store/User/UserTypes';
+import { register } from '@api';
+import Notifs from '@notifications';
+import { NotifTypes } from '@store/Notif/NotifTypes';
+import { popupAlert } from '@components/Alert/Alert.react';
+import { dropdownError } from '@components/Dropdown/Dropdown.react';
+import { setProfileOverride } from '@components/Topbar/Topbar.react';
 import Styles from './Register.style';
 
-type Register3ScreenNavigationProp = StackNavigationProp<
+type RegisterAddressScreenNavigationProp = StackNavigationProp<
   AuthStackParamList,
-  'Register3'
+  'RegisterAddress'
 >;
 
 export interface Props {
-  navigation: Register3ScreenNavigationProp;
+  navigation: RegisterAddressScreenNavigationProp;
   route: {
     params: {
-      firstName: string;
-      lastName: string;
-      referrer: string;
       email: string;
       password: string;
       passwordConfirmation: string;
       remember: boolean;
+      firstName: string;
+      lastName: string;
+      referrer: string;
+      image: Image | undefined;
     };
   };
 }
@@ -39,7 +49,7 @@ export interface State {
   valid: boolean;
 }
 
-class Register3Screen extends React.Component<Props, State> {
+class RegisterAddressScreen extends React.Component<Props, State> {
   private address1 = createRef<Input>();
 
   private address2 = createRef<Input>();
@@ -56,25 +66,34 @@ class Register3Screen extends React.Component<Props, State> {
 
   private unsubscribeFocus: () => void;
 
+  private unsubscribeBlur: () => void;
+
   constructor(props: Props) {
     super(props);
-    this.state = {
-      valid: false,
-    };
     this.onNavigationFocus = this.onNavigationFocus.bind(this);
     this.unsubscribeFocus = this.props.navigation.addListener(
       'focus',
       this.onNavigationFocus
     );
+    this.unsubscribeBlur = this.props.navigation.addListener(
+      'blur',
+      this.onNavigationBlur
+    );
   }
 
   componentWillUnmount(): void {
     this.unsubscribeFocus();
+    this.unsubscribeBlur();
   }
 
   onNavigationFocus(): void {
     if (this.address1.current) this.address1.current.forceFocus();
+    this.updateValid();
   }
+
+  onNavigationBlur = (): void => {
+    setProfileOverride(undefined);
+  };
 
   updateValid = (): void => {
     if (
@@ -90,7 +109,61 @@ class Register3Screen extends React.Component<Props, State> {
         this.city.current.state.valid &&
         this.phyState.current.state.valid &&
         this.postal.current.state.valid;
-      this.setState({ valid: result });
+      setProfileOverride({
+        enabled: result,
+        text: i18n.t('RegisterScreen.register'),
+        action: this.doRegister,
+        blocking: true,
+      });
+    }
+  };
+
+  doRegister = async (): Promise<void> => {
+    Segment.track("Signup - Clicks 'Create Account'");
+    const data: UserRegisterInfo = {
+      ...this.props.route.params,
+      address1: this.address1.current ? this.address1.current.state.value : '',
+      address2: this.address2.current ? this.address2.current.state.value : '',
+      city: this.city.current ? this.city.current.state.value : '',
+      phyState: this.phyState.current ? this.phyState.current.state.value : '',
+      postal: this.postal.current ? this.postal.current.state.value : '',
+    };
+    try {
+      await register(data);
+      Segment.track('Signup - Account Created');
+      Notifs.scheduleNotificationInHours(
+        {
+          title: `${i18n.t('Notifs.youreOneTapAway')}`,
+          body: `${i18n.t('Notifs.clickHereToBegin')}`,
+          data: {
+            type: NotifTypes.NoFirstContact,
+          },
+        },
+        hoursTill8Tomorrow()
+      );
+    } catch (err) {
+      if (err.data && err.data.email) {
+        Segment.trackWithProperties('Signup - Account Creation Error', {
+          'Error Type': 'invalid email',
+        });
+        popupAlert({
+          title: i18n.t('RegisterScreen.emailAlreadyInUse'),
+          buttons: [
+            {
+              text: i18n.t('RegisterScreen.login'),
+              onPress: () => this.props.navigation.replace('Login'),
+            },
+            {
+              text: i18n.t('Alert.okay'),
+              reverse: true,
+            },
+          ],
+        });
+      } else if (err.message === 'timeout') {
+        dropdownError({ message: i18n.t('Error.requestTimedOut') });
+      } else {
+        dropdownError({ message: i18n.t('Error.requestIncomplete') });
+      }
     }
   };
 
@@ -115,7 +188,7 @@ class Register3Screen extends React.Component<Props, State> {
             keyboardShouldPersistTaps="always"
             scrollEnabled
             style={{ width: '100%', flex: 1 }}
-            contentContainerStyle={{ paddingTop: 24, paddingBottom: 72 }}
+            contentContainerStyle={{ paddingVertical: 24 }}
           >
             <Text
               style={[
@@ -135,7 +208,6 @@ class Register3Screen extends React.Component<Props, State> {
               required
               validate={Validation.Address}
               onValid={this.updateValid}
-              onInvalid={() => this.setState({ valid: false })}
               blurOnSubmit={false}
               nextInput={this.address2}
             />
@@ -145,7 +217,6 @@ class Register3Screen extends React.Component<Props, State> {
               placeholder={i18n.t('RegisterScreen.addressLine2')}
               validate={Validation.Address}
               onValid={this.updateValid}
-              onInvalid={() => this.setState({ valid: false })}
               blurOnSubmit={false}
               nextInput={this.city}
             />
@@ -156,7 +227,6 @@ class Register3Screen extends React.Component<Props, State> {
               required
               validate={Validation.City}
               onValid={this.updateValid}
-              onInvalid={() => this.setState({ valid: false })}
               blurOnSubmit={false}
               nextInput={this.phyState}
             />
@@ -174,7 +244,6 @@ class Register3Screen extends React.Component<Props, State> {
                 validate={Validation.State}
                 options={STATES_DROPDOWN}
                 onValid={this.updateValid}
-                onInvalid={() => this.setState({ valid: false })}
                 onFocus={async () => {
                   await sleep(400);
                   if (this.scrollView.current)
@@ -190,73 +259,14 @@ class Register3Screen extends React.Component<Props, State> {
                 required
                 validate={Validation.Postal}
                 onValid={this.updateValid}
-                onInvalid={() => this.setState({ valid: false })}
                 blurOnSubmit={false}
-                onSubmitEditing={() => {
-                  if (this.state.valid) {
-                    this.props.navigation.navigate('Register4', {
-                      ...this.props.route.params,
-                      address1: this.address1.current
-                        ? this.address1.current.state.value
-                        : '',
-                      address2: this.address2.current
-                        ? this.address2.current.state.value
-                        : '',
-                      city: this.city.current
-                        ? this.city.current.state.value
-                        : '',
-                      phyState: this.phyState.current
-                        ? this.phyState.current.state.value
-                        : '',
-                      postal: this.postal.current
-                        ? this.postal.current.state.value
-                        : '',
-                    });
-                  }
-                }}
               />
             </View>
           </ScrollView>
-          <View
-            style={[
-              Styles.fullWidth,
-              {
-                position: 'absolute',
-                bottom: 0,
-                paddingBottom: 8,
-                backgroundColor: 'white',
-              },
-            ]}
-          >
-            <Button
-              containerStyle={[Styles.fullWidth, Styles.registerButton]}
-              buttonText={i18n.t('RegisterScreen.next')}
-              enabled={this.state.valid}
-              onPress={() => {
-                this.props.navigation.navigate('Register4', {
-                  ...this.props.route.params,
-                  address1: this.address1.current
-                    ? this.address1.current.state.value
-                    : '',
-                  address2: this.address2.current
-                    ? this.address2.current.state.value
-                    : '',
-                  city: this.city.current ? this.city.current.state.value : '',
-                  phyState: this.phyState.current
-                    ? this.phyState.current.state.value
-                    : '',
-                  postal: this.postal.current
-                    ? this.postal.current.state.value
-                    : '',
-                });
-              }}
-              showNextIcon
-            />
-          </View>
         </KeyboardAvoidingView>
       </TouchableOpacity>
     );
   }
 }
 
-export default Register3Screen;
+export default RegisterAddressScreen;
