+++ src/views/Onboarding/RegisterPersonal.react.tsx
@@ -1,0 +1,229 @@
+import React, { createRef } from 'react';
+import {
+  KeyboardAvoidingView,
+  ScrollView,
+  Platform,
+  Text,
+  TouchableOpacity,
+  View,
+} from 'react-native';
+import { StackNavigationProp } from '@react-navigation/stack';
+import { AuthStackParamList } from '@navigations';
+import { Button, Input, PicUpload } from '@components';
+import i18n from '@i18n';
+import { Typography } from '@styles';
+import { REFERERS, sleep } from '@utils';
+import { Image } from 'types';
+import { PicUploadTypes } from '@components/PicUpload/PicUpload.react';
+import { setProfileOverride } from '@components/Topbar/Topbar.react';
+import Styles from './Register.style';
+
+type RegisterPersonalScreenNavigationProp = StackNavigationProp<
+  AuthStackParamList,
+  'RegisterPersonal'
+>;
+
+export interface Props {
+  navigation: RegisterPersonalScreenNavigationProp;
+  route: {
+    params: {
+      email: string;
+      password: string;
+      passwordConfirmation: string;
+      remember: boolean;
+    };
+  };
+}
+
+export interface State {
+  valid: boolean;
+  image: Image | undefined;
+}
+
+class RegisterPersonalScreen extends React.Component<Props, State> {
+  private firstName = createRef<Input>();
+
+  private lastName = createRef<Input>();
+
+  private referrer = createRef<Input>();
+
+  private scrollView = createRef<ScrollView>();
+
+  private unsubscribeFocus: () => void;
+
+  private unsubscribeBlur: () => void;
+
+  constructor(props: Props) {
+    super(props);
+    this.state = {
+      valid: false,
+      image: undefined,
+    };
+    this.onNavigationFocus = this.onNavigationFocus.bind(this);
+    this.unsubscribeFocus = this.props.navigation.addListener(
+      'focus',
+      this.onNavigationFocus
+    );
+    this.unsubscribeBlur = this.props.navigation.addListener(
+      'blur',
+      this.onNavigationBlur
+    );
+  }
+
+  componentWillUnmount(): void {
+    this.unsubscribeFocus();
+    this.unsubscribeBlur();
+  }
+
+  onNavigationFocus(): void {
+    if (this.firstName.current) this.firstName.current.forceFocus();
+    this.updateValid();
+  }
+
+  onNavigationBlur = (): void => {
+    setProfileOverride(undefined);
+  };
+
+  goForward = (): void => {
+    this.props.navigation.navigate('RegisterAddress', {
+      ...this.props.route.params,
+      firstName: this.firstName.current
+        ? this.firstName.current.state.value
+        : '',
+      lastName: this.lastName.current ? this.lastName.current.state.value : '',
+      referrer: this.referrer.current ? this.referrer.current.state.value : '',
+      image: this.state.image,
+    });
+  };
+
+  updateValid = (): void => {
+    if (
+      this.firstName.current &&
+      this.lastName.current &&
+      this.referrer.current
+    ) {
+      const result =
+        this.firstName.current.state.valid &&
+        this.lastName.current.state.valid &&
+        this.referrer.current.state.valid;
+      this.setState({ valid: result });
+      setProfileOverride({
+        enabled: result,
+        text: i18n.t('RegisterScreen.next'),
+        action: this.goForward,
+      });
+    }
+  };
+
+  render(): JSX.Element {
+    return (
+      <TouchableOpacity
+        style={{
+          flex: 1,
+          backgroundColor: 'white',
+          paddingHorizontal: 16,
+        }}
+        activeOpacity={1.0}
+      >
+        <KeyboardAvoidingView
+          style={{ flex: 1, flexDirection: 'column', justifyContent: 'center' }}
+          behavior={Platform.OS === 'ios' ? 'padding' : undefined}
+          keyboardVerticalOffset={Platform.OS === 'ios' ? 90 : -200}
+          enabled
+        >
+          <ScrollView
+            ref={this.scrollView}
+            keyboardShouldPersistTaps="always"
+            scrollEnabled
+            style={{ width: '100%', flex: 1 }}
+            contentContainerStyle={{
+              paddingVertical: 24,
+            }}
+          >
+            <TouchableOpacity style={{ flex: 1 }} activeOpacity={1.0}>
+              <Text
+                style={[
+                  Typography.FONT_BOLD,
+                  { fontSize: 20, alignSelf: 'flex-start', paddingBottom: 16 },
+                ]}
+              >
+                {i18n.t('RegisterScreen.registerAccount')}
+              </Text>
+              <View
+                style={{
+                  justifyContent: 'center',
+                  alignItems: 'center',
+                  paddingTop: 24,
+                  width: '100%',
+                  transform: [{ scale: 1.25 }],
+                }}
+              >
+                <PicUpload
+                  type={PicUploadTypes.Profile}
+                  avatarPlaceholder
+                  onSuccess={(image: Image) => this.setState({ image })}
+                  onDelete={() => this.setState({ image: undefined })}
+                />
+              </View>
+              <Text
+                style={[
+                  Typography.FONT_REGULAR,
+                  {
+                    fontSize: 14,
+                    alignSelf: 'center',
+                    paddingBottom: 16,
+                    paddingTop: 32,
+                  },
+                ]}
+              >
+                {i18n.t('RegisterScreen.clickToUploadProfileImage')}
+              </Text>
+              <Input
+                ref={this.firstName}
+                parentStyle={Styles.fullWidth}
+                placeholder={i18n.t('RegisterScreen.firstName')}
+                required
+                onValid={this.updateValid}
+                onInvalid={() => this.setState({ valid: false })}
+                blurOnSubmit={false}
+                nextInput={this.lastName}
+              />
+              <Input
+                ref={this.lastName}
+                parentStyle={Styles.fullWidth}
+                placeholder={i18n.t('RegisterScreen.lastName')}
+                required
+                onValid={this.updateValid}
+                onInvalid={() => this.setState({ valid: false })}
+                blurOnSubmit={false}
+                nextInput={this.referrer}
+              />
+              <Input
+                ref={this.referrer}
+                parentStyle={Styles.fullWidth}
+                placeholder={i18n.t('RegisterScreen.referrer')}
+                required
+                options={REFERERS}
+                onFocus={async () => {
+                  await sleep(400);
+                  if (this.scrollView.current)
+                    this.scrollView.current.scrollToEnd({ animated: true });
+                }}
+                onValid={this.updateValid}
+                onInvalid={() => this.setState({ valid: false })}
+                blurOnSubmit={false}
+                onSubmitEditing={() => {
+                  if (this.state.valid) {
+                    this.goForward();
+                  }
+                }}
+              />
+            </TouchableOpacity>
+          </ScrollView>
+        </KeyboardAvoidingView>
+      </TouchableOpacity>
+    );
+  }
+}
+
+export default RegisterPersonalScreen;
